@@ -3,23 +3,29 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@emotion/react": "11.7.1",
+    "@emotion/react": "11.10.4",
+    "@popperjs/core": "2.11.6",
     "@testing-library/jest-dom": "^5.16.1",
     "@testing-library/react": "^12.1.2",
     "@testing-library/user-event": "^13.5.0",
     "@theme-ui/color": "0.14.5",
     "@types/jest": "^27.0.3",
-    "@types/node": "^16.11.14",
-    "@types/react": "^17.0.37",
-    "@types/react-dom": "^17.0.11",
-    "framer-motion": "5.5.5",
+    "@types/node": "^18.11.3",
+    "@types/react": "18.0.21",
+    "@types/react-dom": "18.0.6",
+    "date-fns": "2.29.3",
+    "framer-motion": "7.6.1",
     "graphql-request": "3.7.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-query": "3.34.5",
-    "react-router-dom": "6.2.1",
-    "react-scripts": "5.0.0",
-    "theme-ui": "0.12.1",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
+    "react-focus-lock": "2.9.1",
+    "react-hook-form": "7.38.0",
+    "react-icons": "4.6.0",
+    "react-popper": "2.3.0",
+    "react-query": "3.39.2",
+    "react-router-dom": "6.4.2",
+    "react-scripts": "5.0.1",
+    "theme-ui": "0.15.3",
     "web-vitals": "^2.1.2"
   },
   "scripts": {
