@@ -1,5 +1,5 @@
 import React from "react";
-import ReactDOM from "react-dom";
+import { createRoot } from "react-dom/client";
 import { QueryClient, QueryClientProvider } from "react-query";
 import { ReactQueryDevtools } from "react-query/devtools";
 import { BrowserRouter as Router } from "react-router-dom";
@@ -14,7 +14,10 @@ import { theme } from "./theme";
 const queryClient = new QueryClient();
 mswInit();
 
-ReactDOM.render(
+const container = document.getElementById("root");
+const root = createRoot(container as HTMLDivElement); // createRoot(container!) if you use TypeScript
+
+root.render(
   <React.StrictMode>
     <QueryClientProvider client={queryClient}>
       <Router>
@@ -32,10 +35,6 @@ ReactDOM.render(
       />
     </QueryClientProvider>
   </React.StrictMode>,
-  document.getElementById("root"),
 );
 
-// If you want to start measuring performance in your app, pass a function
-// to log results (for example: reportWebVitals(console.log))
-// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
 reportWebVitals();
