@@ -3,7 +3,7 @@ import { Card as UICard, CardProps } from "theme-ui";
 
 interface ICard extends CardProps {}
 
-const Card = forwardRef<HTMLDivElement, ICard>(
+export const Card = forwardRef<HTMLDivElement, ICard>(
   ({ children, ...props }, ref) => {
     return (
       <UICard ref={ref} {...props}>
@@ -14,5 +14,3 @@ const Card = forwardRef<HTMLDivElement, ICard>(
 );
 
 Card.displayName = "Card";
-
-export default Card;
