@@ -1,1 +1,1 @@
-export { default } from "./MainContent";
+export * from "./MainContent";
